+++ utils.go
@@ -130,29 +130,35 @@ func removeWhiteSpaces(str string) string {
 
 type ContainerInfo struct {
 	ID     string
+	Name     string
 	Status string
+	ExitCode string
 	Exists bool
+	Logs string
 }
 
 // Returns: container ID, status , whether or not a container exists, error
-func getContainerInfo(shellService ShellServiceInterface, containerNameOrID string) (ContainerInfo, error) {
+func getContainerInfo(shellService ShellServiceInterface, containerNameOrID string) (*ContainerInfo, error) {
 	if containerNameOrID == "" {
 		panic("containerNameOrID was empty")
 	}
-	cmd := fmt.Sprintf("docker inspect --format='{{.Id}} {{.State.Status}}' %s", containerNameOrID)
+	// https://docs.docker.com/engine/api/v1.21/
+	cmd := fmt.Sprintf("docker inspect --format='{{.Id}} {{.Name}} {{.State.Status}} {{.State.ExitCode}}' %s", containerNameOrID)
 	stdout, stderr, exitStatus, _ := shellService.RunGetOutput(cmd, true)
 	if exitStatus != 0 {
 		if strings.Contains(stdout, "No such object") || strings.Contains(stderr, "No such object") {
-			return ContainerInfo{Exists: false}, nil
+			return &ContainerInfo{Exists: false}, nil
 		}
 		cmdInfo := cmdInfoToString(cmd, stdout, stderr, exitStatus)
-		return ContainerInfo{}, fmt.Errorf("Unexpected exit status:\n%s", cmdInfo)
+		return &ContainerInfo{}, fmt.Errorf("Unexpected exit status:\n%s", cmdInfo)
 	}
 	status := strings.TrimSuffix(stdout, "\n")
 	outputArr := strings.Split(status, " ")
-	return ContainerInfo{
+	return &ContainerInfo{
 		ID:     outputArr[0],
-		Status: outputArr[1],
+		Name:     strings.TrimSuffix(outputArr[1], "/"),
+		Status: outputArr[2],
+		ExitCode: outputArr[3],
 		Exists: true,
 	}, nil
 }