+++ utils_test.go
@@ -127,13 +127,15 @@ bb
 func Test_getContainerInfo(t *testing.T) {
 	logger := NewLogger("debug")
 	commandsReactions := make(map[string]interface{}, 0)
-	fakeOutput := `1234 running`
-	commandsReactions["docker inspect --format='{{.Id}} {{.State.Status}}' 1234"] =
+	fakeOutput := `1234 name1 running 133`
+	commandsReactions["docker inspect --format='{{.Id}} {{.Name}} {{.State.Status}} {{.State.ExitCode}}' 1234"] =
 		[]string{fakeOutput, "", "0" }
 	shell := NewMockedShellServiceNotInteractive2(logger, commandsReactions)
 	info, err := getContainerInfo(shell, "1234")
 	assert.Equal(t, "1234", info.ID)
+	assert.Equal(t, "name1", info.Name)
 	assert.Equal(t, "running", info.Status)
+	assert.Equal(t, "133", info.ExitCode)
 	assert.Equal(t, nil, err)
 	assert.Equal(t, true, info.Exists)
 }
@@ -141,7 +143,7 @@ func Test_getContainerInfo(t *testing.T) {
 func Test_getContainerInfo_NoSuchObject(t *testing.T) {
 	logger := NewLogger("debug")
 	commandsReactions := make(map[string]interface{}, 0)
-	commandsReactions["docker inspect --format='{{.Id}} {{.State.Status}}' 1234"] =
+	commandsReactions["docker inspect --format='{{.Id}} {{.Name}} {{.State.Status}} {{.State.ExitCode}}' 1234"] =
 		[]string{"", "Error: No such object: 1234", "1" }
 	shell := NewMockedShellServiceNotInteractive2(logger, commandsReactions)
 	info, err := getContainerInfo(shell, "1234")